allinone/requirements.txt
flask
confluent-kafka
mysql-connector-python